support: add missing_tests.sh and refactor missing_api.sh
* Add missing_tests.sh to measure approximate test coverage using dehydra +
  gcov.
* Refactor missing_api.sh to have a config and a ignore file as arguments.

git-svn-id: https://vcs.maemo.org/svn/pymaemo@1359 f8ba266b-2117-0410-b6ba-c3dd9a426789

+++ support/missing_api.sh
@@ -3,22 +3,30 @@
 # original C API.
 set -e -u
 
+if [ $# -ne 2 ]; then
+    echo "Usage: $0 <config.sh> <ignore.txt>" >&2
+    exit 1
+fi
+base_dir=$(dirname $0)
+config=$1
+ignore=$2
+
+if [ ! -f $ignore ]; then
+    ignore=$base_dir/$ignore
+fi
+if [ ! -f $ignore ]; then
+    echo "Ignore file not found: $ignore" >&2
+    exit 1
+fi
+
 tmp=$(mktemp -d)
 trap "echo Removing temporary files... >&2; rm -rf $tmp" EXIT
 
-base_dir=$(dirname $0)
-# pkg-config module name
-pkgconfig="libosso"
-# extra pkg-config modules for the Python extension
-pkgconfig_extra="pygtk-2.0 gtk+-2.0"
-# extra CFLAGS for the Python extension
-cflags_extra="-I/usr/include/python2.5 -Wno-write-strings -include /usr/include/glib-2.0/glib/gmain.h"
-# command the will return the list of C files for Python API
-#py_api_cmd='find $HOME/workspace/python-osso/python-osso-0.3/src -name "*.c" | egrep -v "/osso-(mime|misc).c$"'
-py_api_cmd='find /tmp/python-osso/python-osso-0.4/osso -name "*.c"'
-# list of C files for C API (usually headers)
-#headers=$(dpkg -L libosso-dev | grep '\.h$')
-headers=/usr/include/libosso.h
+if [ -f $config ]; then
+    . $config
+else
+    . $base_dir/$config
+fi
 
 function dehydra_run()
 {
@@ -35,15 +43,15 @@ function dehydra_run()
 function create_c_plugin()
 {
     # Plugin to collect C API
-    cat > $tmp/from_c.js << EOF
+    cat << EOF
 function process_decl(decl) {
     var headers = new Array();
     var i = 0;
 EOF
     for h in $@; do
-        echo "    headers[i++] = \"$h\";" >> $tmp/from_c.js
+        echo "    headers[i++] = \"$h\";"
     done
-cat >> $tmp/from_c.js << EOF
+    cat << EOF
     for each(h in headers) {
         if (decl.isFunction && decl.loc.file == h)
             print(decl.shortName);
@@ -57,16 +65,16 @@ function create_python_plugin()
     file=$(readlink -f $1); shift
 
     # Plugin to collect Python API
-    cat > $tmp/from_python.js << EOF
+    cat << EOF
 function handle_function_call(fn, stmt)
 {
     var headers = new Array();
     var i = 0;
 EOF
     for h in $@; do
-        echo "    headers[i++] = \"$h\";" >> $tmp/from_python.js
+        echo "    headers[i++] = \"$h\";"
     done
-cat >> $tmp/from_python.js << EOF
+    cat << EOF
     for each(h in headers) {
         if (stmt.loc.file == h && fn.loc.file == "$file")
             _print(stmt.shortName);
@@ -115,53 +123,21 @@ function process(vars, state) {
 EOF
 }
 
-# Ignore list
-cat > $tmp/ignore.txt << EOF
-# - has variable argument list
-# - reimplemented using osso_rpc_run_with_argfill()
-osso_rpc_run
-
-# - has variable argument list
-# - reimplemented using osso_rpc_run_system_with_argfill()
-osso_rpc_run_system
-
-# - has variable argument list
-# - reimplemented using osso_rpc_run_with_argfill()
-osso_rpc_run_with_defaults
-
-# - has variable argument list
-# - reimplemented using osso_rpc_async_run_with_argfill()
-osso_rpc_async_run
-
-# - has variable argument list
-# - reimplemented using osso_rpc_async_run_with_argfill()
-osso_rpc_async_run_with_defaults
-
-# Resource deallocation functions, not necessary in Python
-osso_rpc_free_val
-osso_rpc_set_cb_f_with_free
-osso_rpc_set_default_cb_f_with_free
-
-# Deprecated API
-osso_application_set_top_cb
-osso_application_unset_top_cb
-EOF
-
 # Collect C API
-create_c_plugin $headers
+create_c_plugin $headers > $tmp/from_c.js
 for f in $headers; do
     dehydra_run $tmp/from_c.js $f `pkg-config --cflags $pkgconfig` $cflags_extra
 done | sort -u > $tmp/c_api.txt
 
 # Collect Python API
 eval $py_api_cmd | while read f; do
-    create_python_plugin $f $headers
+    create_python_plugin $f $headers > $tmp/from_python.js
     dehydra_run $tmp/from_python.js $f \
         `pkg-config --cflags $pkgconfig $pkgconfig_extra` $cflags_extra
 done | sort -u > $tmp/called_functions.txt
 
 # List missing API
-egrep -v -e '^$' -e '^#' $tmp/ignore.txt |
+egrep -v -e '^$' -e '^#' $ignore |
 grep -x -F -f $tmp/called_functions.txt -f- -v $tmp/c_api.txt > $tmp/missing.txt
 cat $tmp/missing.txt
